+++ src/utils/space_object.rs
@@ -4,7 +4,7 @@ use crate::utils::vector2::Vector2;
 pub struct SpaceObject {
     pub pos: Vector2,
     pub dir: Vector2,
-    pub angle: f64,
+    pub angle: f32,
     pub radius: usize,
     pub model: Box<[Vector2]>,
 }
